requirements.txt
flask
gradio_client
pillow